% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.peaks.from.gmm.R
\name{.generate.peaks.from.gmm}
\alias{.generate.peaks.from.gmm}
\title{Title}
\usage{
.generate.peaks.from.gmm(dp_data, PARAMETERS, GENEINFO)
}
\arguments{
\item{dp_data}{}

\item{PARAMETERS}{}

\item{GENEINFO}{}
}
\value{

}
\description{
Title
}
