% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.merged.peaks.R
\name{.plot.merged.peaks}
\alias{.plot.merged.peaks}
\title{Title}
\usage{
.plot.merged.peaks(
  plot.startvec,
  plot.fit.frame,
  plot.bin.counts,
  plot.dp,
  plot.merged.peaks,
  PARAMETERS
)
}
\arguments{
\item{plot.startvec}{}

\item{plot.fit.frame}{}

\item{plot.bin.counts}{}

\item{plot.dp}{}

\item{PARAMETERS}{}
}
\value{

}
\description{
Title
}
