% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.peaks.as.granges.R
\name{.retrieve.peaks.as.granges}
\alias{.retrieve.peaks.as.granges}
\title{Title}
\usage{
.retrieve.peaks.as.granges(PEAKS, GENEINFO)
}
\arguments{
\item{PEAKS}{}

\item{GENEINFO}{}
}
\value{

}
\description{
Title
}
