% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPDE4PM.R
\name{DPDE4PM}
\alias{DPDE4PM}
\title{Title}
\usage{
DPDE4PM(
  GENE,
  PEAKS,
  GTF,
  RESOLUTION = 50,
  DP.ITERATIONS = 1000,
  OUTPUTDIR = ".",
  PLOT.RESULT = F
)
}
\arguments{
\item{OUTPUTDIR}{}
}
\value{

}
\description{
Title
}
