% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gene.peaks.R
\name{plot.gene.peaks}
\alias{plot.gene.peaks}
\title{Title}
\usage{
\method{plot}{gene.peaks}(GENE, PEAKS, GTF, OUTPUTDIR = "~/figures")
}
\arguments{
\item{OUTPUTDIR}{}
}
\value{

}
\description{
Title
}
